Web based on flask is build to use the simple word to word replacement. Also LSTM Algorithm based on tensorflow is build to be an alternative way for replacement.
Data: use the whole corpus from nltk sentences, 6_gram to extract the sents that ends with a GRE word. Thus, the 5 words will be the features and the 6th word will be the label. Those 5 words will be convert to a hot map vector [5*lexicon], label will be hot map vector [gre_words].

Model is trained on random 5000 examples and the rest 1000 was used to test (accuracy is 11.2%). Since the train data is not sufficient and we only used 6_gram, and the rnn_size is not as big as the size(lexicon), the accuracy is kind of low. (This is slow enough to train)

Model is saved to do prediction.